move zip/stream methods to IOUtil, fix hidden file logic
