+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/config/XmlJetConfigLoadFromFileSystemOrClasspathTest.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.jet.config;
 
-import com.hazelcast.jet.impl.util.Util;
+import com.hazelcast.jet.impl.util.IOUtil;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import org.junit.runner.RunWith;
 
@@ -39,7 +39,7 @@ public class XmlJetConfigLoadFromFileSystemOrClasspathTest extends AbstractJetCo
         File tempFile = File.createTempFile("jet", ".xml");
         try (FileOutputStream os = new FileOutputStream(tempFile)) {
             InputStream resourceAsStream = getClass().getClassLoader().getResourceAsStream(TEST_XML_JET);
-            os.write(Util.readFully(resourceAsStream));
+            os.write(IOUtil.readFully(resourceAsStream));
         }
 
         // When
