+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/config/XmlJetConfigWithSystemPropertyTest.java
@@ -19,7 +19,7 @@ package com.hazelcast.jet.config;
 import com.hazelcast.core.HazelcastException;
 import com.hazelcast.jet.impl.config.ConfigProvider;
 import com.hazelcast.jet.impl.config.XmlJetConfigBuilder;
-import com.hazelcast.jet.impl.util.Util;
+import com.hazelcast.jet.impl.util.IOUtil;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -62,7 +62,7 @@ public class XmlJetConfigWithSystemPropertyTest extends AbstractJetMemberConfigW
         File tempFile = File.createTempFile("jet", ".xml");
         try (FileOutputStream os = new FileOutputStream(tempFile)) {
             InputStream resourceAsStream = getClass().getClassLoader().getResourceAsStream(TEST_XML_1);
-            os.write(Util.readFully(resourceAsStream));
+            os.write(IOUtil.readFully(resourceAsStream));
         }
         System.setProperty(SYSPROP_JET_CONFIG, tempFile.getAbsolutePath());
 
@@ -107,7 +107,7 @@ public class XmlJetConfigWithSystemPropertyTest extends AbstractJetMemberConfigW
         File tempFile = File.createTempFile("imdg", ".xml");
         try (FileOutputStream os = new FileOutputStream(tempFile)) {
             InputStream resourceAsStream = getClass().getClassLoader().getResourceAsStream(TEST_XML_2);
-            os.write(Util.readFully(resourceAsStream));
+            os.write(IOUtil.readFully(resourceAsStream));
         }
         System.setProperty(SYSPROP_MEMBER_CONFIG, tempFile.getAbsolutePath());
 
@@ -243,7 +243,7 @@ public class XmlJetConfigWithSystemPropertyTest extends AbstractJetMemberConfigW
         File tempFile = File.createTempFile("jet", ".xml");
         try (FileOutputStream os = new FileOutputStream(tempFile)) {
             InputStream resourceAsStream = getClass().getClassLoader().getResourceAsStream("hazelcast-jet-foo.bar");
-            os.write(Util.readFully(resourceAsStream));
+            os.write(IOUtil.readFully(resourceAsStream));
         }
         System.setProperty(SYSPROP_JET_CONFIG, tempFile.getAbsolutePath());
