+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/config/YamlJetClientConfigWithSystemPropertyTest.java
@@ -20,7 +20,7 @@ import com.hazelcast.client.config.ClientConfig;
 import com.hazelcast.client.config.YamlClientConfigBuilder;
 import com.hazelcast.core.HazelcastException;
 import com.hazelcast.jet.impl.config.YamlJetClientConfigLocator;
-import com.hazelcast.jet.impl.util.Util;
+import com.hazelcast.jet.impl.util.IOUtil;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -73,7 +73,7 @@ public class YamlJetClientConfigWithSystemPropertyTest extends AbstractJetConfig
         File tempFile = File.createTempFile("jet", ".yaml");
         try (FileOutputStream os = new FileOutputStream(tempFile)) {
             InputStream resourceAsStream = getClass().getClassLoader().getResourceAsStream(JET_CLIENT_YAML);
-            os.write(Util.readFully(resourceAsStream));
+            os.write(IOUtil.readFully(resourceAsStream));
         }
         System.setProperty(SYSPROP_CLIENT_CONFIG, tempFile.getAbsolutePath());
